requirements.txt
panda
numpy
scikit-learn
seaborn
dill
#-e .